+++ src/lib/escape.js
@@ -8,5 +8,6 @@ export default function escape(str) {
         .replace(/</g, '&lt;')
         .replace(/>/g, '&gt;')
         .replace(/\//g, '&#x2F;')
+        .replace(/\\/g, '&#x5C;')
         .replace(/`/g, '&#96;'));
 }